Update Radar Scope MRA (#41)
The MRA was originally using a mame set of Radar Scope that is now considered an unknown rom, possibly bootlegged or unreliable (now renamed to radarscpc). A reliably sourced rom dump was released which only had minor changes to the xml/driver for it.

Since the previous rom this MRA was pointing to is not considered a known-good dump, then it should be changed. I'm also adding the previous version, known as Rev C now, into the MRA Alternatives in case there is a game breaking regression.